requirements.txt
django-nose==1.3
mock==1.0.1
